Changing over to using doubles instead of floats for better precision

+++ Space.DTO/Buildings/CashFactory.cs
@@ -13,8 +13,8 @@ namespace Space.DTO.Buildings
 
         #region Implementation of ISpatialEntity
 
-        public float Latitude { get; set; }
-        public float Longitude { get; set; }
+        public double Latitude { get; set; }
+        public double Longitude { get; set; }
 
         #endregion
     }
