+++ Space.DTO/Spatial/AsteroidBelt.cs
@@ -16,9 +16,9 @@ namespace Space.DTO.Spatial
 
         #region ISpatialEntity
 
-        public float Latitude { get; set; }
+        public double Latitude { get; set; }
 
-        public float Longitude { get; set; }
+        public double Longitude { get; set; }
 
         #endregion ISpatialEntity
     }
